refactor(landing): extract redirect logic to client component
Move authentication-based redirection logic from the main landing page into a dedicated LandingRedirect component. This change allows app/page.tsx to be rendered as a server component, improving SEO and performance while maintaining the same auth behavior.
